Updated the addTapped function.
Added a constant named context and set it to equal to (UIApplication.shared.delegate as! AppDelegate).persistentContainer.viewContext.

Added a constant named game and set it to equal to Game (context: context).

Set the title value of game to value of the TextField.text.

Set the image value of game to UIImagePNGRepresentation(gameImageView.image!) as NSData?.

Added the following line of code to run the saveContext method, (UIApplication.shared.delegate as! AppDelegate).saveContext().

Created the Game entity into the CoreData.

Added an image attribute and to set it to equal to Binary Data, unchecked Optional and check Allows External Storage.

Added a title attribute and to set it to equal to String and unchecked Optional.