Added the method popToRootViewController of the navigationController to go back to the first view controller of the games list.
Updated the ViewController class type to the class the UITableViewDataSource and UITableViewDelegate.

Set the dateSource of the tableView to self.

Set the delegate of the tableView to self.

Added the method reloadData of the tableView into the viewWillAppear function.

Updated the returned value to games.count for the numberOfRowsInSection function.

Added the cellForRowAt indexPath function.

Added the cell constant and set to equal to UITableViewCell ()

Added the game constant and set to equal to games[indexPath.row]

Set the text value for the textLabel of the cell to the value the game.title.

Set the image value for the imageView of the cell to the value the game.image.

Returned the value of cell.