Added a property variable named game of type Game! (CoreDate Object) and set it to initially equal to nil.
Added an If Statement block to check if game is not equal to nil and print the proper message.

Added the didSelectRowAt indexPath function to obtain the index value for the tapped cell.

Added a constant named game and set it to equal to value of indexPath.row for games.

Added the prepare for segue function.

Added a constant named nextVC equal to segue.destination as! GameViewController.

Set the nextVC.game to equal to sender as? Game.

Set the Identifier of the segue which takes to the Game View Controller to gameSegue.